requirements.txt
langchain
langchain_community
openai
python-dotenv